Fix Update when column is not yet added
This commit fixes an issues when the target model evolves over
several migrations. The statement
    acts_as_archive :indexes => ...
evolves faster than the migrations.
For example, adding an index on a new created column will
work only after the column creation. On the other hand,
some other devs / users may not have run the first migration
yet.
    ActsAsArchive.update Article
will try to add this new index on a non-existing column
and the whole migration process will fail.
We now wrap the index queries within begin/rescue blocks
to avoid failing the whole migration if an index can't be
added or removed